Updated the landing page to a UDF pattern.
This is still going to be in draft but this does work currently I just do not want to miss any functionality and make sure I am coding to the correct standards. There are some areas currently using !! or ? and I think there is a better way to handle it.